backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
azure-cognitiveservices-speech>=1.40.0
python-dotenv>=1.0.0
pymongo>=4.8.0
httpx>=0.27.0
sse-starlette>=2.0.0
